feat(kubectl): Extract interface for KubectlJobExecutor and use it instead of the implementation.
