+++ clouddriver-kubernetes/src/main/java/com/netflix/spinnaker/clouddriver/kubernetes/security/KubernetesCredentials.java
@@ -55,9 +55,9 @@ import com.netflix.spinnaker.clouddriver.kubernetes.description.manifest.Kuberne
 import com.netflix.spinnaker.clouddriver.kubernetes.names.KubernetesNamerRegistry;
 import com.netflix.spinnaker.clouddriver.kubernetes.op.handler.KubernetesCustomResourceHandler;
 import com.netflix.spinnaker.clouddriver.kubernetes.op.handler.KubernetesHandler;
-import com.netflix.spinnaker.clouddriver.kubernetes.op.job.KubectlJobExecutor.KubectlException;
-import com.netflix.spinnaker.clouddriver.kubernetes.op.job.KubectlJobExecutor.KubectlNotFoundException;
-import com.netflix.spinnaker.clouddriver.kubernetes.op.job.KubectlJobExecutorInterface;
+import com.netflix.spinnaker.clouddriver.kubernetes.op.job.DefaultKubectlJobExecutor.KubectlException;
+import com.netflix.spinnaker.clouddriver.kubernetes.op.job.DefaultKubectlJobExecutor.KubectlNotFoundException;
+import com.netflix.spinnaker.clouddriver.kubernetes.op.job.KubectlJobExecutor;
 import com.netflix.spinnaker.kork.configserver.ConfigFileService;
 import com.netflix.spinnaker.moniker.Namer;
 import io.kubernetes.client.openapi.models.V1DeleteOptions;
@@ -91,7 +91,7 @@ public class KubernetesCredentials {
 
   private final Registry registry;
   private final Clock clock;
-  private final KubectlJobExecutorInterface jobExecutor;
+  private final KubectlJobExecutor jobExecutor;
   private final GlobalResourcePropertyRegistry globalResourcePropertyRegistry;
 
   @Include @Getter @Nonnull private final String accountName;
@@ -149,7 +149,7 @@ public class KubernetesCredentials {
 
   public KubernetesCredentials(
       Registry registry,
-      KubectlJobExecutorInterface jobExecutor,
+      KubectlJobExecutor jobExecutor,
       ManagedAccount managedAccount,
       AccountResourcePropertyRegistry.Factory resourcePropertyRegistryFactory,
       KubernetesKindRegistry.Factory kindRegistryFactory,
@@ -806,7 +806,7 @@ public class KubernetesCredentials {
   public static class Factory {
     private final Registry spectatorRegistry;
     private final KubernetesNamerRegistry kubernetesNamerRegistry;
-    private final KubectlJobExecutorInterface jobExecutor;
+    private final KubectlJobExecutor jobExecutor;
     private final ConfigFileService configFileService;
     private final AccountResourcePropertyRegistry.Factory resourcePropertyRegistryFactory;
     private final KubernetesKindRegistry.Factory kindRegistryFactory;
