+++ clouddriver-kubernetes/src/test/java/com/netflix/spinnaker/clouddriver/kubernetes/caching/view/provider/KubernetesDataProviderIntegrationTest.java
@@ -63,7 +63,7 @@ import com.netflix.spinnaker.clouddriver.kubernetes.op.handler.KubernetesReplica
 import com.netflix.spinnaker.clouddriver.kubernetes.op.handler.KubernetesServiceHandler;
 import com.netflix.spinnaker.clouddriver.kubernetes.op.handler.KubernetesUnregisteredCustomResourceHandler;
 import com.netflix.spinnaker.clouddriver.kubernetes.op.handler.ManifestFetcher;
-import com.netflix.spinnaker.clouddriver.kubernetes.op.job.KubectlJobExecutor;
+import com.netflix.spinnaker.clouddriver.kubernetes.op.job.DefaultKubectlJobExecutor;
 import com.netflix.spinnaker.clouddriver.kubernetes.security.*;
 import com.netflix.spinnaker.clouddriver.model.Application;
 import com.netflix.spinnaker.clouddriver.model.HealthState;
@@ -574,8 +574,8 @@ final class KubernetesDataProviderIntegrationTest {
     softly.assertThat(coordinates).isEmpty();
   }
 
-  private static KubectlJobExecutor getJobExecutor() {
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class, new ReturnsSmartNulls());
+  private static DefaultKubectlJobExecutor getJobExecutor() {
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class, new ReturnsSmartNulls());
     when(jobExecutor.list(
             any(KubernetesCredentials.class),
             anyList(),
