+++ clouddriver-kubernetes/src/test/java/com/netflix/spinnaker/clouddriver/kubernetes/security/KubernetesCredentialsTest.java
@@ -44,9 +44,9 @@ import com.netflix.spinnaker.clouddriver.kubernetes.description.manifest.Kuberne
 import com.netflix.spinnaker.clouddriver.kubernetes.names.KubernetesManifestNamer;
 import com.netflix.spinnaker.clouddriver.kubernetes.names.KubernetesNamerRegistry;
 import com.netflix.spinnaker.clouddriver.kubernetes.op.handler.KubernetesUnregisteredCustomResourceHandler;
-import com.netflix.spinnaker.clouddriver.kubernetes.op.job.KubectlJobExecutor;
-import com.netflix.spinnaker.clouddriver.kubernetes.op.job.KubectlJobExecutor.KubectlException;
-import com.netflix.spinnaker.clouddriver.kubernetes.op.job.KubectlJobExecutor.KubectlNotFoundException;
+import com.netflix.spinnaker.clouddriver.kubernetes.op.job.DefaultKubectlJobExecutor;
+import com.netflix.spinnaker.clouddriver.kubernetes.op.job.DefaultKubectlJobExecutor.KubectlException;
+import com.netflix.spinnaker.clouddriver.kubernetes.op.job.DefaultKubectlJobExecutor.KubectlNotFoundException;
 import com.netflix.spinnaker.kork.configserver.CloudConfigResourceService;
 import com.netflix.spinnaker.kork.configserver.ConfigFileService;
 import java.util.HashMap;
@@ -59,7 +59,7 @@ final class KubernetesCredentialsTest {
   private final String OP_NAME = "KubernetesCredentialsTest";
   private final Task task = new DefaultTask("task-id");
 
-  private KubernetesCredentials getCredentials(Registry registry, KubectlJobExecutor jobExecutor) {
+  private KubernetesCredentials getCredentials(Registry registry, DefaultKubectlJobExecutor jobExecutor) {
     KubernetesCredentials.Factory factory =
         new KubernetesCredentials.Factory(
             registry,
@@ -90,7 +90,7 @@ final class KubernetesCredentialsTest {
 
   @Test
   void metricTagsForSuccessfulDeploy() {
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class);
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class);
     Registry registry = new DefaultRegistry();
     KubernetesCredentials credentials = getCredentials(registry, jobExecutor);
     credentials.deploy(getManifest(), task, OP_NAME);
@@ -111,7 +111,7 @@ final class KubernetesCredentialsTest {
 
   @Test
   void metricTagsForSuccessfulList() {
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class);
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class);
     Registry registry = new DefaultRegistry();
     KubernetesCredentials credentials = getCredentials(registry, jobExecutor);
     credentials.list(
@@ -133,7 +133,7 @@ final class KubernetesCredentialsTest {
 
   @Test
   void metricTagsForSuccessfulListNoNamespace() {
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class);
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class);
     Registry registry = new DefaultRegistry();
     KubernetesCredentials credentials = getCredentials(registry, jobExecutor);
     credentials.list(ImmutableList.of(KubernetesKind.DEPLOYMENT, KubernetesKind.REPLICA_SET), null);
@@ -148,7 +148,7 @@ final class KubernetesCredentialsTest {
 
   @Test
   void metricTagsForSuccessfulListEmptyNamespace() {
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class);
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class);
     Registry registry = new DefaultRegistry();
     KubernetesCredentials credentials = getCredentials(registry, jobExecutor);
     credentials.list(ImmutableList.of(KubernetesKind.DEPLOYMENT, KubernetesKind.REPLICA_SET), "");
@@ -163,7 +163,7 @@ final class KubernetesCredentialsTest {
 
   @Test
   void returnValueForSuccessfulList() {
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class);
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class);
     Registry registry = new DefaultRegistry();
     KubernetesCredentials credentials = getCredentials(registry, jobExecutor);
 
@@ -178,7 +178,7 @@ final class KubernetesCredentialsTest {
 
   @Test
   void timeRecordedForSuccessfulList() {
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class);
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class);
 
     ManualClock clock = new ManualClock();
     Registry registry = new DefaultRegistry(clock);
@@ -204,7 +204,7 @@ final class KubernetesCredentialsTest {
 
   @Test
   void metricTagsForListThrowingKubectlException() {
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class);
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class);
     Registry registry = new DefaultRegistry();
     KubernetesCredentials credentials = getCredentials(registry, jobExecutor);
 
@@ -237,7 +237,7 @@ final class KubernetesCredentialsTest {
 
   @Test
   void propagatedExceptionForListThrowingKubectlException() {
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class);
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class);
     Registry registry = new DefaultRegistry();
     KubernetesCredentials credentials = getCredentials(registry, jobExecutor);
 
@@ -257,7 +257,7 @@ final class KubernetesCredentialsTest {
 
   @Test
   void timeRecordedForListThrowingKubectlException() {
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class);
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class);
 
     ManualClock clock = new ManualClock();
     Registry registry = new DefaultRegistry(clock);
@@ -288,7 +288,7 @@ final class KubernetesCredentialsTest {
 
   @Test
   void metricTagsForListThrowingOtherException() {
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class);
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class);
     Registry registry = new DefaultRegistry();
     KubernetesCredentials credentials = getCredentials(registry, jobExecutor);
 
@@ -319,7 +319,7 @@ final class KubernetesCredentialsTest {
 
   @Test
   void timeRecordedForListThrowingOtherException() {
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class);
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class);
 
     ManualClock clock = new ManualClock();
     Registry registry = new DefaultRegistry(clock);
@@ -349,7 +349,7 @@ final class KubernetesCredentialsTest {
 
   @Test
   void propagatedExceptionForListThrowingOtherException() {
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class);
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class);
     Registry registry = new DefaultRegistry();
     KubernetesCredentials credentials = getCredentials(registry, jobExecutor);
 
@@ -369,7 +369,7 @@ final class KubernetesCredentialsTest {
   @Test
   void replaceWhenResourceExists() {
     KubernetesManifest manifest = getManifest();
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class);
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class);
     KubernetesCredentials credentials = getCredentials(new NoopRegistry(), jobExecutor);
     when(jobExecutor.create(credentials, manifest, task, OP_NAME))
         .thenThrow(new KubectlException("Create failed: Error from server (AlreadyExists)"));
@@ -382,7 +382,7 @@ final class KubernetesCredentialsTest {
   @Test
   void replaceWhenResourceDoesNotExist() {
     KubernetesManifest manifest = getManifest();
-    KubectlJobExecutor jobExecutor = mock(KubectlJobExecutor.class);
+    DefaultKubectlJobExecutor jobExecutor = mock(DefaultKubectlJobExecutor.class);
     KubernetesCredentials credentials = getCredentials(new NoopRegistry(), jobExecutor);
     when(jobExecutor.replace(credentials, manifest, task, OP_NAME))
         .thenThrow(new KubectlNotFoundException("Not found"));
